requirements.txt
gunicorn
pandas==0.20.3
Flask==0.12.3
scikit-learn==0.19.0
requests==2.18.4
matplotlib==2.0.2
scipy==0.19.1
numpy
xlrd==1.1.0
urllib3==1.22
cloudant
python-docx==0.7.4
Flask-BasicAuth
